Added event, event_callback, event_dispatcher
The event API is modeled somewhat after the ActionScript 3 events API. There is no capture phase, and bubbling is not yet implemented. The major addition is
generic listeners, which will trigger on any event. Generic listeners can be
added with an event_type of :*, or using the add_generic_listener utility
method; likewise for removing generic listeners.