Implemented event bubbling
To enable event bubbling, define a @parent ivar, or set the __bubble_attr__ to a
valid string or symbol. Valid symbols include :@@foo to use a cvar, :@bar to use
an ivar, or :baz to use an instance method. Remember to set event[:bubbles] =
true, or dispatch_event event, :bubbles => :true.