Added ActionDelegate and DelegateController
The DelegateController is intended to allow module builders to delegate action
handling to other objects. For example, a controller for exploring a physical
space might delegate actions to the object representing that space, such as the
:go action or the :look action. The space object might further delegate some
actions, such as :talk if there are any people to talk to in that space.